Refactor token command
