+++ gitlab/kubectl.go
@@ -1,0 +1,12 @@
+package gitlab
+
+import (
+	"github.com/appscode/go/term"
+	"github.com/skratchdot/open-golang/open"
+)
+
+func IssueToken() {
+	codeURurl := "https://gitlab.com/profile/personal_access_tokens"
+	term.Infoln("Gitlab url for personal access tokens:", codeURurl)
+	open.Start(codeURurl)
+}
